最初に表示されるページを AppRouter に移動
