--- pages/_document.tsx
@@ -1,19 +1,0 @@
-import NextDocument, { Html, Head, Main, NextScript } from 'next/document';
-
-export default class Document extends NextDocument {
-    render() {
-        return (
-            <Html>
-                <Head>
-                    {/* PWA */}
-                    <link rel="icon" sizes="192x192" href="/icon.png" />
-                    <link rel="manifest" href="/manifest.json" />
-                </Head>
-                <body>
-                    <Main />
-                    <NextScript />
-                </body>
-            </Html>
-        );
-    }
-}
